# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Historial Clinico Backend",
        "description": "Aplicación backend del sistema web Historial Clnico",
        "contact": {
            "email": "dav.almeyda@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/historial-clinico-backend/public/api/obs/tratamiento": {
            "post": {
                "tags": [
                    "Observaciones"
                ],
                "summary": "Obtener la observación de un tratamiento",
                "operationId": "tratamiento",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal",
                                    "numAtendMed"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "numAtendMed": {
                                        "type": "string",
                                        "example": "0000384457"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
...
        "/historial-clinico-backend/public/api/tratamientos/antecedentes": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener lista de antecedentes de tratamientos",
                "operationId": "tratamientosAntecedentes",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codPaciente",
                                    "codMedico"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codPaciente": {
                                        "type": "string",
                                        "example": "0010185756"
                                    },
                                    "codMedico": {
                                        "type": "string",
                                        "example": "0000026144"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/tratamientos/estomatologico": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener lista de tratamientos estomatologico",
                "operationId": "tratamientosEstomatologico",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codPaciente",
                                    "codMedico",
                                    "nroAtencion"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codPaciente": {
                                        "type": "string",
                                        "example": "0010185756"
                                    },
                                    "codMedico": {
                                        "type": "string",
                                        "example": "0000026144"
                                    },
                                    "nroAtencion": {
                                        "type": "string",
                                        "example": "0000347913"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/tratamientos": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener tratamientos",
                "operationId": "tratamientos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/tratamientos/recetaSugerido": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener receta sugerida de tratamientos",
                "operationId": "tratamientosRecetaSugerido",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codProdRuc"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codProdRuc": {
                                        "type": "string",
                                        "example": "000023@20603070349"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/tratamientos/getRecomendaciones": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener recomendaciones de tratamientos",
                "operationId": "tratamientosGetRecomendaciones",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nunReceta",
                                    "atencionMedica",
                                    "codProducto"
                                ],
                                "properties": {
                                    "nunReceta": {
                                        "type": "string"
                                    },
                                    "atencionMedica": {
                                        "type": "string"
                                    },
                                    "codProducto": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/tratamientos/setRecomendaciones": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Guardar recomendaciones de tratamientos",
                "operationId": "tratamientosSetRecomendaciones",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nunReceta",
                                    "atencionMedica",
                                    "codProducto",
                                    "recomendacion"
                                ],
                                "properties": {
                                    "nunReceta": {
                                        "type": "string"
                                    },
                                    "atencionMedica": {
                                        "type": "string"
                                    },
                                    "codProducto": {
                                        "type": "string"
                                    },
                                    "recomendacion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}